fix(rio): enable deletion of old deploy files
There's an issue with things like PathPlanner where if you don't clean the deploy firectory on the RIO, it can lead to some orphaned files that can mess up the code. I believe this also has the added benefit of forcing the update of edited files, which was a problem that you had to SSH into the RIO to fix. It was annoying. I don't want to have that happen this year.